Endret andel: flytt ut logikk som genererer tidslinje pr pers
